Move `IntoAgents` to `stateful`
