+++ packages/engine/lib/stateful/src/agent/mod.rs
@@ -1,8 +1,10 @@
 // TODO: DOC: Add module level docs for describing the high level concept of agents
 
 mod field;
+mod into_agent;
 mod name;
 mod references;
+mod required;
 mod schema;
 
 // temporarily public
@@ -11,7 +13,9 @@ pub mod arrow;
 pub use self::{
     arrow::{flush_pending_columns, modify_loaded_column, AgentBatch, AgentPool},
     field::{Agent, AgentStateField, BUILTIN_FIELDS},
+    into_agent::IntoAgents,
     name::AgentName,
     references::MessageReference,
+    required::IsRequired,
     schema::AgentSchema,
 };
