+++ packages/engine/src/datastore/arrow/mod.rs
@@ -1,4 +1,2 @@
-pub mod batch_conversion;
-
 #[cfg(test)]
 mod test;
