+++ packages/engine/src/datastore/table/create_remove/command.rs
@@ -1,13 +1,10 @@
 use std::{collections::HashSet, sync::Arc};
 
 use arrow::record_batch::RecordBatch;
-use stateful::agent::AgentSchema;
+use stateful::{agent::AgentSchema, field::UUID_V4_LEN};
 
 use crate::{
-    datastore::{
-        error::{Error, Result},
-        UUID_V4_LEN,
-    },
+    datastore::error::{Error, Result},
     simulation::command::CreateRemoveCommands,
 };
