+++ packages/engine/src/simulation/package/context/packages/agent_messages/collected.rs
@@ -1,10 +1,7 @@
-use stateful::message::MessageMap;
+use stateful::{field::UUID_V4_LEN, message::MessageMap};
 
-use crate::{
-    datastore::UUID_V4_LEN,
-    simulation::package::context::packages::agent_messages::{
-        indices::AgentMessageIndices, Result,
-    },
+use crate::simulation::package::context::packages::agent_messages::{
+    indices::AgentMessageIndices, Result,
 };
 
 /// Columnar native representation of indices to messages
