+++ packages/engine/src/simulation/package/context/packages/api_requests/handlers.rs
@@ -1,13 +1,9 @@
 use serde_json::Value;
-use stateful::agent;
+use stateful::{agent, field::UUID_V4_LEN, message::MessageReader};
 use thiserror::Error as ThisError;
-use stateful::message::MessageReader;
 
-use crate::{
-    datastore::UUID_V4_LEN,
-    simulation::package::context::packages::api_requests::{
-        handlers, ApiResponseMap, Error, Result,
-    },
+use crate::simulation::package::context::packages::api_requests::{
+    handlers, ApiResponseMap, Error, Result,
 };
 
 pub const ACTIVE_REQUESTS: usize = 10;
@@ -63,14 +59,12 @@ pub mod mapbox {
     use std::collections::{hash_map, HashMap};
 
     use futures::StreamExt;
+    use stateful::field::UUID_V4_LEN;
     use thiserror::Error as ThisError;
 
-    use crate::{
-        datastore::UUID_V4_LEN,
-        simulation::package::context::packages::api_requests::{
-            handlers::{CustomError, Request, Requests, ACTIVE_REQUESTS},
-            ApiResponseMap, Result, Value,
-        },
+    use crate::simulation::package::context::packages::api_requests::{
+        handlers::{CustomError, Request, Requests, ACTIVE_REQUESTS},
+        ApiResponseMap, Result, Value,
     };
 
     #[derive(ThisError, Debug)]
